+++ pkgs/intl4x/lib/src/datetime_format/datetime_format_impl.dart
@@ -96,12 +96,23 @@ abstract class FormatterStandaloneImpl extends DateTimeFormatterStandalone {
 
   String formatInternal(DateTime datetime);
 
+  String formatRangeInternal(DateTime start, DateTime end);
+
   @override
   String format(DateTime datetime) {
     if (isInTest) {
       return '$datetime//${_impl.locale}';
     } else {
       return formatInternal(datetime);
+    }
+  }
+
+  @override
+  String formatRange(DateTime start, DateTime end) {
+    if (isInTest) {
+      return '$start//$end//${_impl.locale}';
+    } else {
+      return formatRangeInternal(start, end);
     }
   }
 }
@@ -147,6 +158,10 @@ sealed class DateTimeFormatterStandalone {
   /// Formats the given [datetime] into a string according to the formatter's
   /// configured locale and options.
   String format(DateTime datetime);
+
+  /// Formats the given date range from [start] to [end] into a string according
+  /// to the formatter's configured locale and options.
+  String formatRange(DateTime start, DateTime end);
 }
 
 /// Formatters that can format a [DateTime] with time zone information.