+++ pkgs/intl4x/lib/src/datetime_format/icu4x/date_time_formatter.dart
@@ -11,6 +11,7 @@ import 'datetime_format_4x.dart';
 /// Wraps an [icu.DateTimeFormatter]
 class DateTimeFormatterX extends FormatterImpl {
   final icu.DateTimeFormatter formatter;
+  final icu.DateTimeRangeFormatterGregorian rangeFormatter;
   final DateTimeFormatImpl impl;
   final icu.Locale localeX;
 
@@ -26,6 +27,12 @@ class DateTimeFormatterX extends FormatterImpl {
         length: length ?? icu.DateTimeLength.short,
         timePrecision: timePrecision,
       ),
+      rangeFormatter = icu.DateTimeRangeFormatterGregorian.mdt(
+        localeX,
+        alignment: alignment,
+        length: length ?? icu.DateTimeLength.short,
+        timePrecision: timePrecision,
+      ),
       super(impl);
 
   DateTimeFormatterX.ymdt(
@@ -36,6 +43,13 @@ class DateTimeFormatterX extends FormatterImpl {
     icu.TimePrecision? timePrecision,
     icu.YearStyle? yearStyle,
   ) : formatter = icu.DateTimeFormatter.ymdt(
+        localeX,
+        alignment: alignment,
+        length: length ?? icu.DateTimeLength.short,
+        timePrecision: timePrecision,
+        yearStyle: yearStyle,
+      ),
+      rangeFormatter = icu.DateTimeRangeFormatterGregorian.ymdt(
         localeX,
         alignment: alignment,
         length: length ?? icu.DateTimeLength.short,
@@ -58,12 +72,31 @@ class DateTimeFormatterX extends FormatterImpl {
         timePrecision: timePrecision,
         yearStyle: yearStyle,
       ),
+      rangeFormatter = icu.DateTimeRangeFormatterGregorian.ymdet(
+        localeX,
+        alignment: alignment,
+        length: length ?? icu.DateTimeLength.short,
+        timePrecision: timePrecision,
+        yearStyle: yearStyle,
+      ),
       super(impl);
 
   @override
   String formatInternal(DateTime datetime) {
     final (isoDate, time) = datetime.toX;
     return formatter.formatIso(isoDate, time);
+  }
+
+  @override
+  String formatRangeInternal(DateTime start, DateTime end) {
+    final (startIsoDate, startTime) = start.toX;
+    final (endIsoDate, endTime) = end.toX;
+    return rangeFormatter.formatIso(
+      startIsoDate,
+      startTime,
+      endIsoDate,
+      endTime,
+    );
   }
 
   @override